+++ feed/index.html
@@ -67,7 +67,7 @@
                         <h1 class="mb-3 mt-3">Latest Posts</h1>
                         <div class="select-container"></div>
                         <select class="form-select mb-3" aria-label="Default select example">
-                            <option selected>Filter by category </option>
+                            <option selected value="show-all-posts">All posts</option>
                         </select>
                         <div class="postsContainer row justify-content-center">
 
@@ -118,8 +118,7 @@
         crossorigin="anonymous"></script>
 
     <script src="/js/components/logOutButton.mjs" type="module"></script>
-    <script src="/js/fetchPosts.js" type="module"></script>
-    <script src="/js/fetchPostsWithTag.mjs" type="module"></script>
+    <script src="/js/setupTagsAndPosts.mjs" type="module"></script>
 
 </body>
