--- js/fetchPostsWithTag.mjs
@@ -1,3 +1,0 @@
-import { API_BASE_URL } from "./settings/baseUrl.mjs";
-import { token } from "./utils/storage.mjs";
-import { fetchAllPosts, postsURL } from "./fetchPosts.js";
